requirements.txt
flask==3.0.0
pillow==10.1.0
torch==2.3.0
diffusers==0.27.0
transformers==4.38.0
accelerate==0.27.0